Convert nodeCIDR to switch to use existing network+subnet (#63)
CAPO uses nodeCIDR as a switch. When it is set, CAPO creates a network
and subnet. When it is not set, it uses existing a network and a subnet,
which are must declared by filters.

See https://github.com/kubernetes-sigs/cluster-api-provider-openstack/blob/73890460b8074ab55b3e3cb2545a2dd6f4a49739/controllers/openstackcluster_controller.go#L382

Co-authored-by: Paweł Kopiczko <kopiczko@gmail.com>