Support labelling MachineDeployments
Labels applied at the MachineDeployment level are taken into
consideration by cluster-autoscaler, i.e if you have multiple node pools
and you're using a nodeSelector in your deployment to target nodes with
a particular label.  If the right label exists for the
MachineDeployment, then cluster-autoscaler knows which pool to scale.